Add files via upload
Project Title:  Portfolio Website

Description:
This repository hosts my portfolio website, a project created from scratch using HTML, CSS, and JavaScript. The purpose of this website is to serve as a platform for me to showcase my work, skills, and accomplishments.

Key Features:

HTML Structure: The website's foundation is built with clean and semantic HTML, ensuring an organized and accessible structure.

CSS Styling: Cascading Style Sheets have been meticulously used to create an appealing and responsive design that enhances the user experience.

JavaScript Interactivity: JavaScript has been employed to add interactivity, making the website engaging and user-friendly. This includes interactive elements, animations, and other dynamic features.

Usage:
Feel free to explore the live version of this website [insert the website link here] to view my work and get to know me better. It's not just a showcase of my skills but also a representation of my dedication to web development and design.

Contributions:
I welcome contributions and suggestions for improvements. If you find any issues or have ideas to enhance this website, please create a pull request or open an issue.

Future Plans:
I have plans to expand this portfolio by adding more projects and blog posts in the future to keep it current and informative.

Acknowledgments:
I'd like to express my gratitude to the open-source community and the creators of the technologies I've used in this project. Your work has been instrumental in making this portfolio possible.